Removed ControlValidationMixin mixin
