+++ addon/components/bs4/bs-form/element/control/input.js
@@ -1,7 +1,9 @@
 import FormElementControlInput from 'ember-bootstrap/components/base/bs-form/element/control/input';
-import ControlValidationMixin from 'ember-bootstrap/mixins/control-validation';
 import SizeClass from 'ember-bootstrap/mixins/size-class';
+import formValidationClass from 'ember-bootstrap/utils/form-validation-class';
 
-export default FormElementControlInput.extend(ControlValidationMixin, SizeClass, {
-  classTypePrefix: 'form-control'
+export default FormElementControlInput.extend(SizeClass, {
+  classTypePrefix: 'form-control',
+  classNameBindings: ['formValidationClass'],
+  formValidationClass: formValidationClass()
 });
